Feature/#178 estimate concurrent

+++ src/main/java/com/ureca/request/application/RequestService.java
@@ -13,6 +13,7 @@ import com.ureca.request.domain.Request;
 import com.ureca.request.infrastructure.RequestRepository;
 import com.ureca.request.presentation.dto.RequestDto;
 import com.ureca.review.domain.Enum.AuthorType;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -95,6 +96,32 @@ public class RequestService {
 
         Customer customer = pet.getCustomer(); // Pet에 Customer 연관 관계가 있어야 합니다.
         String region = customer.getAddress1() + customer.getDetailAddress();
+        BigDecimal deliveryFee = BigDecimal.ZERO;
+        if (Boolean.TRUE.equals(requestDto.getIsVisitRequired())) {
+            // 견종의 종류에 따라 가격을 설정
+            switch (pet.getMajorBreedCode()) {
+                case "P1":
+                    deliveryFee = BigDecimal.valueOf(20000);
+                    break;
+                case "P2":
+                    deliveryFee = BigDecimal.valueOf(30000);
+                    break;
+                case "P3":
+                    deliveryFee = BigDecimal.valueOf(40000);
+                    break;
+                case "P4":
+                    deliveryFee = BigDecimal.valueOf(50000);
+                    break;
+                default:
+                    throw new ApiException(ErrorCode.INVALID_BREED);
+            }
+        }
+
+        // isMonitoringIncluded에 따라 monitoringFee 계산
+        BigDecimal monitoringFee =
+                Boolean.TRUE.equals(requestDto.getIsMonitoringIncluded())
+                        ? BigDecimal.valueOf(50000)
+                        : BigDecimal.ZERO;
         Request request =
                 Request.builder()
                         .pet(pet)
@@ -110,6 +137,8 @@ public class RequestService {
                         .isDelivery(requestDto.getIsVisitRequired())
                         .isMonitoringIncluded(requestDto.getIsMonitoringIncluded())
                         .additionalRequest(requestDto.getAdditionalRequest())
+                        .deliveryFee(deliveryFee)
+                        .monitoringFee(monitoringFee)
                         .requestStatus("ST1") // 기본 상태
                         .build();
 
@@ -180,6 +209,8 @@ public class RequestService {
                 .isVisitRequired(request.getIsDelivery())
                 .isMonitoringIncluded(request.getIsMonitoringIncluded())
                 .additionalRequest(request.getAdditionalRequest())
+                .deliveryFee(request.getDeliveryFee())
+                .monitoringFee(request.getMonitoringFee())
                 .build();
     }
 
@@ -217,7 +248,11 @@ public class RequestService {
                 requestRepository
                         .findById(requestId)
                         .orElseThrow(() -> new ApiException(ErrorCode.REQUEST_NOT_EXIST));
-        requestRepository.delete(request);
+        if (request.getRequestStatus() == "ST1" || request.getRequestStatus() == "ST3") {
+            requestRepository.delete(request);
+        } else {
+            throw new ApiException(ErrorCode.RESERVE_EXIST_ERROR);
+        }
     }
 
     public List<RequestDto.Response> selectDesignerRequest(Long designerId) {
